moved the --run-relevance option + “skip relevance tests by default” logic into a packaged pytest plugin that is included in the wheel and auto-loaded by pytest.
